requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
requests>=2.28.0
python-dotenv>=0.19.0
scikit-learn>=1.3.0
nltk>=3.8